rmf_ai_dreams/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
chromadb>=0.4.0
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0
mistralai>=0.0.11
selenium>=4.15.0
playwright>=1.40.0
python-dotenv>=1.0.0
cryptography>=41.0.0
Pillow>=10.1.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
